Refactor lambda quickfixes for headless
- ChangeLambdaBodyToBlock
- ChangeLambdaBodyToExpression
- ConvertLambdaToMethodReference
- AddInferredLambdaParameterTypes
- AddVarLambdaParameterTypes
- RemoveVarOrInferredLambdaParameterTypes
